added baseUrl + retrofit DI

+++ app/src/main/java/com/emperor/hpproject/data/network/models/HPWand.kt
@@ -1,3 +1,3 @@
 package com.emperor.hpproject.data.network.models
 
-data class HPWand(val wood: String, val core: String, val length: Int?)
+data class HPWand(val wood: String, val core: String, val length: Float?)
